webserv: logger ref 2
