+++ srcs/main.cpp
@@ -1,14 +1,16 @@
 #include <iostream>
 
-#include "Server.h"
+//#include "Server.h"
+#include "Logger/Logger.hpp"
 
 int main(int ac, char **av)
 {
+    Logger  log("DEBUG", "webserv.log");
     (void)av;
     if (ac < 2)
     {
-        std::cerr << "Using default configuration" << std::endl;
+        log.write(INFO, "Config file not set. Using default configuration");
     }
-    Server webserv;
-    webserv.Run();
+    /* Server webserv;
+    webserv.Run();*/
 }
